refactor: move annotation queue code to ee folder
