+++ web/src/ee/features/annotation-queues/components/AnnotationQueueItemPage.tsx
