+++ web/src/ee/features/annotation-queues/components/DeleteAnnotationQueueButton.tsx
